// LiveAuction/src/type.ts
export interface AucStateResType {
  auc_title: string;
  auc_kind: number;
  auc_num: number;
  auc_date: string;
  auc_now_date: string;
  lot_stat_cd: string;
  lot_row_num: number;
  lot_total_cnt: number;
  auc_place: string;
  auc_start_date: string;
  bid_increase_price: number;
  auc_end_date: string;
  my_paddle_num: number;
  my_bid_type: number;
  auc_stat_cd: string;
  img_file_name: string;
  lot_num: number;
  pre_bid_price: number;
  bid_price: number;
  next_bid_price: number;
  successful_bid: boolean;
  successful_bid_price: number;
  successful_paddle_num: number;
  pre_bid_proc_yn: boolean;
  is_fair_warning: boolean;
  usd_next_bid_price: number;
  jpy_next_bid_price: number;
  cny_next_bid_price: number;
  hkd_next_bid_price: number;
  eur_next_bid_price: number;
  usd_bid_price: number;
  jpy_bid_price: number;
  cny_bid_price: number;
  hkd_bid_price: number;
  eur_bid_price: number;
  claim_bid_hst_seq: number;
}

// 작품리스트 조회
export interface AuctionWorkResponse {
  w_make_date: string;
  img_file_name: string | undefined;
  t_seq: number;
  lot_num: number;
  a_name: string;
  w_name: string;
  w_price: number;
  w_low_price: number;
  w_high_price: number;
  usd_w_low_price: number;
  usd_w_high_price: number;
  jpy_w_low_price: number;
  jpy_w_high_price: number;
  cny_w_low_price: number;
  cny_w_high_price: number;
  hkd_w_low_price: number;
  hkd_w_high_price: number;
  eur_w_low_price: number;
  eur_w_high_price: number;
  w_unit: string;
  w_size: string;
  successful_bid_price: number;
  successful_bid: boolean; // 낙찰여부. false가 낙찰
  is_fair_warning: boolean;
  work_seq: number;
  isWish: boolean;
}